+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0
 
 Latest
 ------
+* Update Observation admin to show checklist date, time, location and observer.
 
 1.0.0 (2025-02-14)
 ------------------